+++ src/LocalDateTime.php
@@ -685,7 +685,7 @@ final class LocalDateTime implements JsonSerializable
      */
     public function toDateTime(): DateTime
     {
-        return $this->atTimeZone(TimeZone::utc())->toDateTime();
+        return $this->atTimeZone(TimeZone::utc())->toNativeDateTime();
     }
 
     /**